+++ intg-appletv/config.py
@@ -17,7 +17,7 @@ from typing import Iterator
 
 import discover
 import pyatv
-from ucapi import Entity, EntityTypes
+from ucapi import EntityTypes
 
 _LOG = logging.getLogger(__name__)
 
@@ -29,15 +29,6 @@ class AtvProtocol(str, Enum):
 
     AIRPLAY = "airplay"
     COMPANION = "companion"
-
-
-class AppleTVEntity(Entity):
-    """Global AppleTV entity."""
-
-    @property
-    def deviceid(self) -> str:
-        """Return the device identifier."""
-        raise NotImplementedError()
 
 
 @dataclass